+++ src/extra/layouts/grid/lv_grid.c
@@ -102,7 +102,7 @@ static inline uint8_t get_grid_row_align(lv_obj_t * obj)
 /**********************
  *  GLOBAL VARIABLES
  **********************/
-uint32_t LV_LAYOUT_GRID;
+uint16_t LV_LAYOUT_GRID;
 lv_style_prop_t LV_STYLE_GRID_COLUMN_DSC_ARRAY;
 lv_style_prop_t LV_STYLE_GRID_COLUMN_ALIGN;
 lv_style_prop_t LV_STYLE_GRID_ROW_DSC_ARRAY;
